+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `TagInputAutocomplete`: Add component.
+- Feature: `Autocomplete`: Move to `input` dir and modernize component.
+
 ## [10.4.1] - 2022-04-28
 
 Fixes: